Split repo.ts and search.ts
