+++ packages/server/src/fhir/repo.ts
@@ -4,45 +4,30 @@ import {
   canReadResourceType,
   canWriteResourceType,
   deepEquals,
-  DEFAULT_SEARCH_COUNT,
   evalFhirPath,
-  evalFhirPathTyped,
   experimentalValidateResource,
-  FhirFilterComparison,
-  FhirFilterConnective,
-  FhirFilterExpression,
-  FhirFilterNegation,
-  Operator as FhirOperator,
-  Filter,
   forbidden,
   formatSearchQuery,
-  getReferenceString,
   getSearchParameterDetails,
   getStatus,
   gone,
-  IncludeTarget,
   isGone,
   isNotFound,
   isOk,
-  isResource,
   matchesAccessPolicy,
   normalizeErrorString,
   normalizeOperationOutcome,
   notFound,
   OperationOutcomeError,
   parseCriteriaAsSearchRequest,
-  parseFilterParameter,
-  PropertyType,
   protectedResourceTypes,
   publicResourceTypes,
   resolveId,
   SearchParameterDetails,
   SearchParameterType,
   SearchRequest,
-  SortRule,
   stringify,
   tooManyRequests,
-  toTypedValue,
   validateResource,
   validateResourceType,
 } from '@medplum/core';
@@ -52,7 +37,6 @@ import {
   AccessPolicyResource,
   Bundle,
   BundleEntry,
-  BundleLink,
   Meta,
   OperationOutcome,
   Reference,
@@ -64,7 +48,6 @@ import { randomUUID } from 'crypto';
 import { Pool, PoolClient } from 'pg';
 import { applyPatch, Operation } from 'rfc6902';
 import validator from 'validator';
-import { getConfig } from '../config';
 import { getClient } from '../database';
 import { logger } from '../logger';
 import { getRedis } from '../redis';
@@ -88,24 +71,13 @@ import { AddressTable } from './lookups/address';
 import { HumanNameTable } from './lookups/humanname';
 import { LookupTable } from './lookups/lookuptable';
 import { TokenTable } from './lookups/token';
-import { deriveIdentifierSearchParameter } from './lookups/util';
 import { ValueSetElementTable } from './lookups/valuesetelement';
 import { getPatient } from './patient';
 import { validateReferences } from './references';
 import { rewriteAttachments, RewriteMode } from './rewrite';
-import {
-  Column,
-  Condition,
-  Conjunction,
-  DeleteQuery,
-  Disjunction,
-  Expression,
-  InsertQuery,
-  Negation,
-  Operator,
-  SelectQuery,
-} from './sql';
-import { getSearchParameter, getSearchParameters } from './structure';
+import { buildSearchExpression, getFullUrl, searchImpl } from './search';
+import { Condition, DeleteQuery, Disjunction, Expression, InsertQuery, Operator, SelectQuery } from './sql';
+import { getSearchParameters } from './structure';
 
 /**
  * The RepositoryContext interface defines standard metadata for repository actions.
@@ -191,10 +163,10 @@ const lookupTables: LookupTable<unknown>[] = [
   new ValueSetElementTable(),
 ];
 
-/**
- * Defines the maximum number of resources returned in a single search result.
- */
-const maxSearchResults = 1000;
+// /**
+//  * Defines the maximum number of resources returned in a single search result.
+//  */
+// const maxSearchResults = 1000;
 
 /**
  * The Repository class manages reading and writing to the FHIR repository.
@@ -410,7 +382,7 @@ export class Repository extends BaseRepository implements FhirRepository {
               ],
             };
         entries.push({
-          fullUrl: this.getFullUrl(resourceType, row.id),
+          fullUrl: getFullUrl(resourceType, row.id),
           request: {
             method: 'GET',
             url: `${resourceType}/${row.id}/_history/${row.versionId}`,
@@ -925,7 +897,7 @@ export class Repository extends BaseRepository implements FhirRepository {
       .execute(getClient());
   }
 
-  async search<T extends Resource>(searchRequest: SearchRequest): Promise<Bundle<T>> {
+  async search<T extends Resource>(searchRequest: SearchRequest<T>): Promise<Bundle<T>> {
     try {
       const resourceType = searchRequest.resourceType;
       validateResourceType(resourceType);
@@ -934,35 +906,9 @@ export class Repository extends BaseRepository implements FhirRepository {
         throw new OperationOutcomeError(forbidden);
       }
 
-      // Ensure that "count" is set.
-      // Count is an optional field.  From this point on, it is safe to assume it is a number.
-      if (searchRequest.count === undefined) {
-        searchRequest.count = DEFAULT_SEARCH_COUNT;
-      } else if (searchRequest.count > maxSearchResults) {
-        searchRequest.count = maxSearchResults;
-      }
-
-      let entry = undefined;
-      let hasMore = false;
-      if (searchRequest.count > 0) {
-        ({ entry, hasMore } = await this.getSearchEntries<T>(searchRequest));
-      }
-
-      let total = undefined;
-      if (searchRequest.total === 'accurate') {
-        total = await this.getAccurateCount(searchRequest);
-      } else if (searchRequest.total === 'estimate') {
-        total = await this.getEstimateCount(searchRequest);
-      }
-
+      const result = await searchImpl(this, searchRequest);
       this.logEvent(SearchInteraction, AuditEventOutcome.Success, undefined, undefined, searchRequest);
-      return {
-        resourceType: 'Bundle',
-        type: 'searchset',
-        entry,
-        total,
-        link: this.getSearchLinks(searchRequest, hasMore),
-      };
+      return result;
     } catch (err) {
       this.logEvent(SearchInteraction, AuditEventOutcome.MinorFailure, err, undefined, searchRequest);
       throw err;
@@ -970,298 +916,10 @@ export class Repository extends BaseRepository implements FhirRepository {
   }
 
   /**
-   * Returns the bundle entries for a search request.
-   * @param searchRequest The search request.
-   * @returns The bundle entries for the search result.
-   */
-  private async getSearchEntries<T extends Resource>(
-    searchRequest: SearchRequest
-  ): Promise<{ entry: BundleEntry<T>[]; hasMore: boolean }> {
-    const resourceType = searchRequest.resourceType;
-    const client = getClient();
-    const builder = new SelectQuery(resourceType)
-      .column({ tableName: resourceType, columnName: 'id' })
-      .column({ tableName: resourceType, columnName: 'content' });
-
-    this.addSortRules(builder, searchRequest);
-    this.addDeletedFilter(builder);
-    this.addSecurityFilters(builder, resourceType);
-    this.addSearchFilters(builder, searchRequest);
-
-    if (builder.joins.length > 0) {
-      builder.groupBy({ tableName: resourceType, columnName: 'id' });
-    }
-
-    const count = searchRequest.count as number;
-    builder.limit(count + 1); // Request one extra to test if there are more results
-    builder.offset(searchRequest.offset || 0);
-
-    const rows = await builder.execute(client);
-    const resources = rows.slice(0, count).map((row) => JSON.parse(row.content as string)) as T[];
-    const entries = resources.map(
-      (resource) =>
-        ({
-          fullUrl: this.getFullUrl(resourceType, resource.id as string),
-          resource,
-        } as BundleEntry)
-    );
-
-    if (searchRequest.include || searchRequest.revInclude) {
-      let base = resources;
-      let iterateOnly = false;
-      const seen: Record<string, boolean> = {};
-      resources.forEach((r) => {
-        seen[`${r.resourceType}/${r.id}`] = true;
-      });
-      let depth = 0;
-      while (base.length > 0) {
-        // Circuit breaker / load limit
-        if (depth >= 5 || entries.length > 1000) {
-          throw new Error(
-            `Search with _(rev)include reached query scope limit: depth=${depth}, results=${entries.length}`
-          );
-        }
-
-        const includes =
-          searchRequest.include
-            ?.filter((param) => !iterateOnly || param.modifier === FhirOperator.ITERATE)
-            ?.map((param) => this.getSearchIncludeEntries(param, base)) || [];
-        const revincludes =
-          searchRequest.revInclude
-            ?.filter((param) => !iterateOnly || param.modifier === FhirOperator.ITERATE)
-            ?.map((param) => this.getSearchRevIncludeEntries(param, base)) || [];
-
-        const includedResources = (await Promise.all([...includes, ...revincludes])).flat();
-        includedResources.forEach((r) => {
-          const ref = `${r.resource?.resourceType}/${r.resource?.id}`;
-          if (!seen[ref]) {
-            entries.push(r);
-          }
-          seen[ref] = true;
-        });
-        base = includedResources.map((entry) => entry.resource) as T[];
-        iterateOnly = true; // Only consider :iterate params on iterations after the first
-        depth++;
-      }
-    }
-
-    for (const entry of entries) {
-      this.removeHiddenFields(entry.resource as Resource);
-    }
-
-    return {
-      entry: entries as BundleEntry<T>[],
-      hasMore: rows.length > count,
-    };
-  }
-
-  /**
-   * Returns bundle entries for the resources that are included in the search result.
-   *
-   * See documentation on _include: https://hl7.org/fhir/R4/search.html#include
-   * @param include The include parameter.
-   * @param resources The base search result resources.
-   * @returns The bundle entries for the included resources.
-   */
-  private async getSearchIncludeEntries(include: IncludeTarget, resources: Resource[]): Promise<BundleEntry[]> {
-    const searchParam = getSearchParameter(include.resourceType, include.searchParam);
-    if (!searchParam) {
-      throw new OperationOutcomeError(
-        badRequest(`Invalid include parameter: ${include.resourceType}:${include.searchParam}`)
-      );
-    }
-
-    const fhirPathResult = evalFhirPathTyped(searchParam.expression as string, resources.map(toTypedValue));
-
-    const references = fhirPathResult
-      .filter((typedValue) => (typedValue.type as PropertyType) === PropertyType.Reference)
-      .map((typedValue) => typedValue.value as Reference);
-    const readResult = await this.readReferences(references);
-    const includedResources = readResult.filter((e) => isResource(e as Resource | undefined)) as Resource[];
-
-    const canonicalReferences = fhirPathResult
-      .filter((typedValue) => [PropertyType.canonical, PropertyType.uri].includes(typedValue.type as PropertyType))
-      .map((typedValue) => typedValue.value as string);
-    if (canonicalReferences.length > 0) {
-      const canonicalSearches = (searchParam.target || []).map((resourceType) =>
-        this.searchResources({
-          resourceType: resourceType,
-          filters: [
-            {
-              code: 'url',
-              operator: FhirOperator.EQUALS,
-              value: canonicalReferences.join(','),
-            },
-          ],
-        })
-      );
-      (await Promise.all(canonicalSearches)).forEach((resources) => {
-        includedResources.push(...resources);
-      });
-    }
-
-    return includedResources.map(
-      (resource: Resource) =>
-        ({
-          fullUrl: this.getFullUrl(resource.resourceType, resource.id as string),
-          resource,
-        } as BundleEntry)
-    ) as BundleEntry[];
-  }
-
-  /**
-   * Returns bundle entries for the resources that are reverse included in the search result.
-   *
-   * See documentation on _revinclude: https://hl7.org/fhir/R4/search.html#revinclude
-   * @param revInclude The revInclude parameter.
-   * @param resources The base search result resources.
-   * @returns The bundle entries for the reverse included resources.
-   */
-  private async getSearchRevIncludeEntries(revInclude: IncludeTarget, resources: Resource[]): Promise<BundleEntry[]> {
-    const searchParam = getSearchParameter(revInclude.resourceType, revInclude.searchParam);
-    if (!searchParam) {
-      throw new OperationOutcomeError(
-        badRequest(`Invalid include parameter: ${revInclude.resourceType}:${revInclude.searchParam}`)
-      );
-    }
-
-    const paramDetails = getSearchParameterDetails(revInclude.resourceType, searchParam);
-    let value: string;
-    if (paramDetails.type === SearchParameterType.CANONICAL) {
-      value = resources
-        .map((r) => (r as any).url)
-        .filter((u) => u !== undefined)
-        .join(',');
-    } else {
-      value = resources.map(getReferenceString).join(',');
-    }
-
-    return (
-      await this.getSearchEntries({
-        resourceType: revInclude.resourceType as ResourceType,
-        filters: [
-          {
-            code: revInclude.searchParam,
-            operator: FhirOperator.EQUALS,
-            value: value,
-          },
-        ],
-      })
-    ).entry;
-  }
-
-  /**
-   * Returns the search bundle links for a search request.
-   * At minimum, the 'self' link will be returned.
-   * If "count" does not equal zero, then 'first', 'next', and 'previous' links will be included.
-   * @param searchRequest The search request.
-   * @param hasMore True if there are more entries after the current page.
-   * @returns The search bundle links.
-   */
-  private getSearchLinks(searchRequest: SearchRequest, hasMore: boolean | undefined): BundleLink[] {
-    const result: BundleLink[] = [
-      {
-        relation: 'self',
-        url: this.getSearchUrl(searchRequest),
-      },
-    ];
-
-    const count = searchRequest.count as number;
-    if (count > 0) {
-      const offset = searchRequest.offset || 0;
-
-      result.push({
-        relation: 'first',
-        url: this.getSearchUrl({ ...searchRequest, offset: 0 }),
-      });
-
-      if (hasMore) {
-        result.push({
-          relation: 'next',
-          url: this.getSearchUrl({ ...searchRequest, offset: offset + count }),
-        });
-      }
-
-      if (offset > 0) {
-        result.push({
-          relation: 'previous',
-          url: this.getSearchUrl({ ...searchRequest, offset: offset - count }),
-        });
-      }
-    }
-
-    return result;
-  }
-
-  private getFullUrl(resourceType: string, id: string): string {
-    return `${getConfig().baseUrl}fhir/R4/${resourceType}/${id}`;
-  }
-
-  private getSearchUrl(searchRequest: SearchRequest): string {
-    return `${getConfig().baseUrl}fhir/R4/${searchRequest.resourceType}${formatSearchQuery(searchRequest)}`;
-  }
-
-  /**
-   * Returns the total number of matching results for a search request.
-   * This ignores page number and page size.
-   * @param searchRequest The search request.
-   * @returns The total number of matching results.
-   */
-  private async getAccurateCount(searchRequest: SearchRequest): Promise<number> {
-    const client = getClient();
-    const builder = new SelectQuery(searchRequest.resourceType);
-    this.addDeletedFilter(builder);
-    this.addSecurityFilters(builder, searchRequest.resourceType);
-    this.addSearchFilters(builder, searchRequest);
-
-    if (builder.joins.length > 0) {
-      builder.raw(`COUNT (DISTINCT "${searchRequest.resourceType}"."id")::int AS "count"`);
-    } else {
-      builder.raw('COUNT("id")::int AS "count"');
-    }
-
-    const rows = await builder.execute(client);
-    return rows[0].count as number;
-  }
-
-  /**
-   * Returns the estimated number of matching results for a search request.
-   * This ignores page number and page size.
-   * This uses the estimated row count technique as described here: https://wiki.postgresql.org/wiki/Count_estimate
-   * @param searchRequest The search request.
-   * @returns The total number of matching results.
-   */
-  private async getEstimateCount(searchRequest: SearchRequest): Promise<number> {
-    const resourceType = searchRequest.resourceType;
-    const client = getClient();
-    const builder = new SelectQuery(resourceType).column('id');
-    this.addDeletedFilter(builder);
-    this.addSecurityFilters(builder, searchRequest.resourceType);
-    this.addSearchFilters(builder, searchRequest);
-    builder.explain = true;
-
-    if (builder.joins.length > 0) {
-      builder.groupBy({ tableName: resourceType, columnName: 'id' });
-    }
-
-    // See: https://wiki.postgresql.org/wiki/Count_estimate
-    // This parses the query plan to find the estimated number of rows.
-    const rows = await builder.execute(client);
-    for (const row of rows) {
-      const queryPlan = row['QUERY PLAN'];
-      const match = /rows=(\d+)/.exec(queryPlan);
-      if (match) {
-        return parseInt(match[1], 10);
-      }
-    }
-    return 0;
-  }
-
-  /**
    * Adds filters to ignore soft-deleted resources.
    * @param builder The select query builder.
    */
-  private addDeletedFilter(builder: SelectQuery): void {
+  addDeletedFilter(builder: SelectQuery): void {
     builder.where('deleted', Operator.EQUALS, false);
   }
 
@@ -1270,7 +928,7 @@ export class Repository extends BaseRepository implements FhirRepository {
    * @param builder The select query builder.
    * @param resourceType The resource type for compartments.
    */
-  private addSecurityFilters(builder: SelectQuery, resourceType: string): void {
+  addSecurityFilters(builder: SelectQuery, resourceType: string): void {
     if (publicResourceTypes.includes(resourceType)) {
       // No compartment restrictions for public resources.
       return;
@@ -1317,7 +975,7 @@ export class Repository extends BaseRepository implements FhirRepository {
         } else if (policy.criteria) {
           // Add subquery for access policy criteria.
           const searchRequest = parseCriteriaAsSearchRequest(policy.criteria);
-          const accessPolicyExpression = this.buildSearchExpression(builder, searchRequest);
+          const accessPolicyExpression = buildSearchExpression(builder, searchRequest);
           if (accessPolicyExpression) {
             expressions.push(accessPolicyExpression);
           }
@@ -1331,349 +989,6 @@ export class Repository extends BaseRepository implements FhirRepository {
     if (expressions.length > 0) {
       builder.predicate.expressions.push(new Disjunction(expressions));
     }
-  }
-
-  /**
-   * Adds all search filters as "WHERE" clauses to the query builder.
-   * @param selectQuery The select query builder.
-   * @param searchRequest The search request.
-   */
-  private addSearchFilters(selectQuery: SelectQuery, searchRequest: SearchRequest): void {
-    const expr = this.buildSearchExpression(selectQuery, searchRequest);
-    if (expr) {
-      selectQuery.predicate.expressions.push(expr);
-    }
-  }
-
-  private buildSearchExpression(selectQuery: SelectQuery, searchRequest: SearchRequest): Expression | undefined {
-    const expressions: Expression[] = [];
-    if (searchRequest.filters) {
-      for (const filter of searchRequest.filters) {
-        const expr = this.buildSearchFilterExpression(selectQuery, searchRequest, filter);
-        if (expr) {
-          expressions.push(expr);
-        }
-      }
-    }
-    if (expressions.length === 0) {
-      return undefined;
-    }
-    if (expressions.length === 1) {
-      return expressions[0];
-    }
-    return new Conjunction(expressions);
-  }
-
-  /**
-   * Builds a single search filter as "WHERE" clause to the query builder.
-   * @param selectQuery The select query builder.
-   * @param searchRequest The search request.
-   * @param filter The search filter.
-   * @returns The search query where expression
-   */
-  private buildSearchFilterExpression(
-    selectQuery: SelectQuery,
-    searchRequest: SearchRequest,
-    filter: Filter
-  ): Expression | undefined {
-    if (typeof filter.value !== 'string') {
-      throw new OperationOutcomeError(badRequest('Search filter value must be a string'));
-    }
-
-    const specialParamExpression = this.trySpecialSearchParameter(selectQuery, searchRequest, filter);
-    if (specialParamExpression) {
-      return specialParamExpression;
-    }
-
-    const resourceType = searchRequest.resourceType;
-    let param = getSearchParameter(resourceType, filter.code);
-    if (!param?.code) {
-      throw new OperationOutcomeError(badRequest(`Unknown search parameter: ${filter.code}`));
-    }
-
-    if (filter.operator === FhirOperator.IDENTIFIER) {
-      param = deriveIdentifierSearchParameter(param);
-      filter = {
-        ...filter,
-        code: param.code as string,
-        operator: FhirOperator.EQUALS,
-      };
-    }
-
-    const lookupTable = this.getLookupTable(resourceType, param);
-    if (lookupTable) {
-      return lookupTable.buildWhere(selectQuery, resourceType, filter);
-    }
-
-    // Not any special cases, just a normal search parameter.
-    return this.buildNormalSearchFilterExpression(resourceType, param, filter);
-  }
-
-  /**
-   * Builds a search filter expression for a normal search parameter.
-   *
-   * Not any special cases, just a normal search parameter.
-   * @param resourceType The FHIR resource type.
-   * @param param The FHIR search parameter.
-   * @param filter The search filter.
-   * @returns A SQL "WHERE" clause expression.
-   */
-  private buildNormalSearchFilterExpression(resourceType: string, param: SearchParameter, filter: Filter): Expression {
-    const details = getSearchParameterDetails(resourceType, param);
-    if (filter.operator === FhirOperator.MISSING) {
-      return new Condition(details.columnName, filter.value === 'true' ? Operator.EQUALS : Operator.NOT_EQUALS, null);
-    } else if (param.type === 'string') {
-      return this.buildStringSearchFilter(details, filter);
-    } else if (param.type === 'token' || param.type === 'uri') {
-      return this.buildTokenSearchFilter(resourceType, details, filter);
-    } else if (param.type === 'reference') {
-      return this.buildReferenceSearchFilter(details, filter);
-    } else if (param.type === 'date') {
-      return this.buildDateSearchFilter(details, filter);
-    } else if (param.type === 'quantity') {
-      return new Condition(details.columnName, fhirOperatorToSqlOperator(filter.operator), parseFloat(filter.value));
-    } else {
-      return new Condition(details.columnName, fhirOperatorToSqlOperator(filter.operator), filter.value);
-    }
-  }
-
-  /**
-   * Returns true if the search parameter code is a special search parameter.
-   *
-   * See: https://www.hl7.org/fhir/search.html#all
-   * @param selectQuery The select query builder.
-   * @param searchRequest The overall search request.
-   * @param filter The search filter.
-   * @returns True if the search parameter is a special code.
-   */
-  private trySpecialSearchParameter(
-    selectQuery: SelectQuery,
-    searchRequest: SearchRequest,
-    filter: Filter
-  ): Expression | undefined {
-    const resourceType = searchRequest.resourceType;
-    const code = filter.code;
-
-    if (code === '_id') {
-      return this.buildIdSearchFilter(resourceType, { columnName: 'id', type: SearchParameterType.UUID }, filter);
-    }
-
-    if (code === '_lastUpdated') {
-      return this.buildDateSearchFilter({ type: SearchParameterType.DATETIME, columnName: 'lastUpdated' }, filter);
-    }
-
-    if (code === '_compartment' || code === '_project') {
-      return this.buildIdSearchFilter(
-        resourceType,
-        { columnName: 'compartments', type: SearchParameterType.UUID, array: true },
-        filter
-      );
-    }
-
-    if (code === '_filter') {
-      return this.buildFilterParameterExpression(selectQuery, searchRequest, parseFilterParameter(filter.value));
-    }
-
-    return undefined;
-  }
-
-  private buildFilterParameterExpression(
-    selectQuery: SelectQuery,
-    searchRequest: SearchRequest,
-    filterExpression: FhirFilterExpression
-  ): Expression {
-    if (filterExpression instanceof FhirFilterNegation) {
-      return this.buildFilterParameterNegation(selectQuery, searchRequest, filterExpression);
-    } else if (filterExpression instanceof FhirFilterConnective) {
-      return this.buildFilterParameterConnective(selectQuery, searchRequest, filterExpression);
-    } else if (filterExpression instanceof FhirFilterComparison) {
-      return this.buildFilterParameterComparison(selectQuery, searchRequest, filterExpression);
-    } else {
-      throw new OperationOutcomeError(badRequest('Unknown filter expression type'));
-    }
-  }
-
-  private buildFilterParameterNegation(
-    selectQuery: SelectQuery,
-    searchRequest: SearchRequest,
-    filterNegation: FhirFilterNegation
-  ): Expression {
-    return new Negation(this.buildFilterParameterExpression(selectQuery, searchRequest, filterNegation.child));
-  }
-
-  private buildFilterParameterConnective(
-    selectQuery: SelectQuery,
-    searchRequest: SearchRequest,
-    filterConnective: FhirFilterConnective
-  ): Expression {
-    const expressions = [
-      this.buildFilterParameterExpression(selectQuery, searchRequest, filterConnective.left),
-      this.buildFilterParameterExpression(selectQuery, searchRequest, filterConnective.right),
-    ];
-    return filterConnective.keyword === 'and' ? new Conjunction(expressions) : new Disjunction(expressions);
-  }
-
-  private buildFilterParameterComparison(
-    selectQuery: SelectQuery,
-    searchRequest: SearchRequest,
-    filterComparison: FhirFilterComparison
-  ): Expression {
-    return this.buildSearchFilterExpression(selectQuery, searchRequest, {
-      code: filterComparison.path,
-      operator: filterComparison.operator as FhirOperator,
-      value: filterComparison.value,
-    }) as Expression;
-  }
-
-  /**
-   * Adds a string search filter as "WHERE" clause to the query builder.
-   * @param details The search parameter details.
-   * @param filter The search filter.
-   * @returns The select query condition.
-   */
-  private buildStringSearchFilter(details: SearchParameterDetails, filter: Filter): Expression {
-    if (filter.operator === FhirOperator.EXACT) {
-      return new Condition(details.columnName, Operator.EQUALS, filter.value);
-    }
-    return new Condition(details.columnName, Operator.LIKE, '%' + filter.value + '%');
-  }
-
-  /**
-   * Adds an ID search filter as "WHERE" clause to the query builder.
-   * @param resourceType The resource type.
-   * @param details The search parameter details.
-   * @param filter The search filter.
-   * @returns The select query condition.
-   */
-  private buildIdSearchFilter(resourceType: string, details: SearchParameterDetails, filter: Filter): Expression {
-    const column = new Column(resourceType, details.columnName);
-    const expressions = [];
-    for (const valueStr of filter.value.split(',')) {
-      let value: string | boolean = valueStr;
-      if (valueStr.includes('/')) {
-        value = valueStr.split('/').pop() as string;
-      }
-      if (!validator.isUUID(value)) {
-        value = '00000000-0000-0000-0000-000000000000';
-      }
-      if (details.array) {
-        expressions.push(new Condition(column, Operator.ARRAY_CONTAINS, value, details.type + '[]'));
-      } else {
-        expressions.push(new Condition(column, Operator.EQUALS, value));
-      }
-    }
-    const disjunction = new Disjunction(expressions);
-    if (filter.operator === FhirOperator.NOT_EQUALS || filter.operator === FhirOperator.NOT) {
-      return new Negation(disjunction);
-    }
-    return disjunction;
-  }
-
-  /**
-   * Adds a token search filter as "WHERE" clause to the query builder.
-   * @param resourceType The resource type.
-   * @param details The search parameter details.
-   * @param filter The search filter.
-   * @returns The select query condition.
-   */
-  private buildTokenSearchFilter(resourceType: string, details: SearchParameterDetails, filter: Filter): Expression {
-    const column = new Column(resourceType, details.columnName);
-    const expressions = [];
-    for (const valueStr of filter.value.split(',')) {
-      let value: string | boolean = valueStr;
-      if (details.type === SearchParameterType.BOOLEAN) {
-        value = valueStr === 'true';
-      } else if (valueStr.includes('|')) {
-        value = valueStr.split('|').pop() as string;
-      }
-      if (details.array) {
-        expressions.push(new Condition(column, Operator.ARRAY_CONTAINS, value, details.type + '[]'));
-      } else if (filter.operator === FhirOperator.CONTAINS) {
-        expressions.push(new Condition(column, Operator.LIKE, '%' + value + '%'));
-      } else {
-        expressions.push(new Condition(column, Operator.EQUALS, value));
-      }
-    }
-    const disjunction = new Disjunction(expressions);
-    if (filter.operator === FhirOperator.NOT_EQUALS || filter.operator === FhirOperator.NOT) {
-      return new Negation(disjunction);
-    }
-    return disjunction;
-  }
-
-  /**
-   * Adds a reference search filter as "WHERE" clause to the query builder.
-   * @param details The search parameter details.
-   * @param filter The search filter.
-   * @returns The select query condition.
-   */
-  private buildReferenceSearchFilter(details: SearchParameterDetails, filter: Filter): Expression {
-    const values = [];
-    for (const value of filter.value.split(',')) {
-      if (!value.includes('/') && (details.columnName === 'subject' || details.columnName === 'patient')) {
-        values.push('Patient/' + value);
-      } else {
-        values.push(value);
-      }
-    }
-    if (details.array) {
-      return new Condition(details.columnName, Operator.ARRAY_CONTAINS, values);
-    }
-    if (values.length === 1) {
-      return new Condition(details.columnName, Operator.EQUALS, values[0]);
-    }
-    return new Condition(details.columnName, Operator.IN, values);
-  }
-
-  /**
-   * Adds a date or date/time search filter.
-   * @param details The search parameter details.
-   * @param filter The search filter.
-   * @returns The select query condition.
-   */
-  private buildDateSearchFilter(details: SearchParameterDetails, filter: Filter): Expression {
-    const dateValue = new Date(filter.value);
-    if (isNaN(dateValue.getTime())) {
-      throw new OperationOutcomeError(badRequest(`Invalid date value: ${filter.value}`));
-    }
-    return new Condition(details.columnName, fhirOperatorToSqlOperator(filter.operator), filter.value);
-  }
-
-  /**
-   * Adds all "order by" clauses to the query builder.
-   * @param builder The client query builder.
-   * @param searchRequest The search request.
-   */
-  private addSortRules(builder: SelectQuery, searchRequest: SearchRequest): void {
-    searchRequest.sortRules?.forEach((sortRule) => this.addOrderByClause(builder, searchRequest, sortRule));
-  }
-
-  /**
-   * Adds a single "order by" clause to the query builder.
-   * @param builder The client query builder.
-   * @param searchRequest The search request.
-   * @param sortRule The sort rule.
-   */
-  private addOrderByClause(builder: SelectQuery, searchRequest: SearchRequest, sortRule: SortRule): void {
-    if (sortRule.code === '_lastUpdated') {
-      builder.orderBy('lastUpdated', !!sortRule.descending);
-      return;
-    }
-
-    const resourceType = searchRequest.resourceType;
-    const param = getSearchParameter(resourceType, sortRule.code);
-    if (!param?.code) {
-      return;
-    }
-
-    const lookupTable = this.getLookupTable(resourceType, param);
-    if (lookupTable) {
-      lookupTable.addOrderBy(builder, resourceType, sortRule);
-      return;
-    }
-
-    const details = getSearchParameterDetails(resourceType, param);
-    builder.orderBy(details.columnName, !!sortRule.descending);
   }
 
   /**
@@ -2162,7 +1477,7 @@ export class Repository extends BaseRepository implements FhirRepository {
    * @param resourceType The resource type.
    * @returns True if the current user can read the specified resource type.
    */
-  private canReadResourceType(resourceType: string): boolean {
+  canReadResourceType(resourceType: string): boolean {
     if (this.isSuperAdmin()) {
       return true;
     }
@@ -2250,7 +1565,7 @@ export class Repository extends BaseRepository implements FhirRepository {
    * @param input The input resource.
    * @returns The resource with hidden fields removed.
    */
-  private removeHiddenFields<T extends Resource>(input: T): T {
+  removeHiddenFields<T extends Resource>(input: T): T {
     const policy = this.getResourceAccessPolicy(input.resourceType);
     if (policy?.hiddenFields) {
       for (const field of policy.hiddenFields) {
@@ -2371,6 +1686,19 @@ export class Repository extends BaseRepository implements FhirRepository {
   }
 }
 
+export function isIndexTable(resourceType: string, searchParam: SearchParameter): boolean {
+  return !!getLookupTable(resourceType, searchParam);
+}
+
+export function getLookupTable(resourceType: string, searchParam: SearchParameter): LookupTable<unknown> | undefined {
+  for (const lookupTable of lookupTables) {
+    if (lookupTable.isIndexed(searchParam, resourceType)) {
+      return lookupTable;
+    }
+  }
+  return undefined;
+}
+
 /**
  * Tries to read a cache entry from Redis by resource type and ID.
  * @param resourceType The resource type.
@@ -2441,35 +1769,6 @@ async function deleteCacheEntries(resourceType: string, ids: string[]): Promise<
  */
 function getCacheKey(resourceType: string, id: string): string {
   return `${resourceType}/${id}`;
-}
-
-/**
- * Converts a FHIR search operator into a SQL operator.
- * Only works for simple conversions.
- * For complex conversions, need to build custom SQL.
- * @param fhirOperator The FHIR operator.
- * @returns The equivalent SQL operator.
- */
-function fhirOperatorToSqlOperator(fhirOperator: FhirOperator): Operator {
-  switch (fhirOperator) {
-    case FhirOperator.EQUALS:
-      return Operator.EQUALS;
-    case FhirOperator.NOT:
-    case FhirOperator.NOT_EQUALS:
-      return Operator.NOT_EQUALS;
-    case FhirOperator.GREATER_THAN:
-    case FhirOperator.STARTS_AFTER:
-      return Operator.GREATER_THAN;
-    case FhirOperator.GREATER_THAN_OR_EQUALS:
-      return Operator.GREATER_THAN_OR_EQUALS;
-    case FhirOperator.LESS_THAN:
-    case FhirOperator.ENDS_BEFORE:
-      return Operator.LESS_THAN;
-    case FhirOperator.LESS_THAN_OR_EQUALS:
-      return Operator.LESS_THAN_OR_EQUALS;
-    default:
-      throw new Error(`Unknown FHIR operator: ${fhirOperator}`);
-  }
 }
 
 export const systemRepo = new Repository({
